Remove Numpy as a dependency
Numpy was included primarily as a means of initializing
multidimensional arrays. Given this is not hard to do
with native python, numpy has been removed in favor of
initializing the code ourselves.

This removes an unnecessary dependency from our codebase
and should make the application lighter to install and
run.